[Feat] diary 테이블 설정

+++ src/main/java/zerobase/weather/repository/DiaryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import zerobase.weather.domain.Diary;
 
 @Repository
-public interface DiaryRepository extends JpaRepository<Diary, Integer> {
+public interface DiaryRepository extends JpaRepository<Diary, Long> {
 }
